backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.2.2
pydantic==2.12.5
pydantic-settings==2.10.1
langchain==1.3.0
langchain-core==1.4.0
langchain-openai==1.2.1
httpx==0.28.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
passlib==1.7.4
google-auth==2.48.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-generativeai==0.4.0
openai==2.36.0
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
docx2txt==0.8
youtube-transcript-api==0.6.1
requests==2.32.5
beautifulsoup4==4.12.2
pillow==10.1.0
anyio==3.7.1
